html/semantics/sections/the-h1-h2-h3-h4-h5-and-h6-elements/sections-h1-thru-6-elements.html

<!DOCTYPE html>
<html>
<head>
	<meta charset="utf-8">
	<title>H1 - H6 elements</title>
	<link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
	<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/single-page.html#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>

	<script>

	var strElement = "", testElement = {}, testElementProperties = [];

	for (var i = 1; i < 7; i++) {

		strElement = "h" + i;
		
		// check that prototype matches spec's DOM interface
		test(function () {
			testElement = document.createElement(strElement);
			assert_equals(Object.getPrototypeOf(testElement), HTMLHeadingElement.prototype, "HTMLHeadingElement.prototype should be used for " + strElement);
		}, "The prototype for " + strElement + " is HTMLHeadingElement.prototype");

		// not checking sectioning effect of element
	}
	</script>

</head>
<body>
	<h1>Description</h1>
	<p>This test validates the H1 - H6 elements.</p>
	
	<div id="log"></div>
	
</body>
</html>
